revert type change for custom sum

+++ allreduce.cu
@@ -76,10 +76,10 @@ void initializeData(bfloat16 *data, int64_t size) {
     }
 }
 
-void custom_bf16_sum(void *invec, void *inoutvec, int64_t *len, MPI_Datatype *datatype) {
+void custom_bf16_sum(void *invec, void *inoutvec, int *len, MPI_Datatype *datatype) {
     bfloat16* in = (bfloat16*) invec;
     bfloat16* inout = (bfloat16*) inoutvec;
-    for (int64_t i = 0; i < *len; i++) {
+    for (int i = 0; i < *len; i++) {
         #ifdef USE_CUDA
         inout[i] = __hadd(in[i], inout[i]);
         #elif USE_ROCM
